+++ client/components/List-View/index.ts
@@ -1,0 +1,1 @@
+export { ListViewDisplay } from './ListViewDisplay';
