+++ client/components/index.ts
@@ -1,5 +1,5 @@
 export { NodeModal } from './NodeModal';
-export {Node} from './Node';
+export { Node } from './Node';
 export { Tooltip } from './Tooltip';
 export { ThemeContainer } from './ThemeContainer';
 export { Graph } from './Graph';
@@ -7,6 +7,7 @@ export { Draggable } from './Draggable';
 export { SwitchButton } from './SwitchButton';
 export { Notification } from './Notification';
 export { NotificationDisplay } from './NotificationDisplay';
-export { Pod} from './Pod';
+export { Pod } from './Pod';
 export { PodContainer } from './PodContainer';
-export {Container} from './Container';
+export { Container } from './Container';
+export { ListViewDisplay } from './List-View';
