+++ client/index.tsx
@@ -7,30 +7,28 @@ import './css/variables.scss';
 import { store } from './store/store';
 
 const container = document.getElementById('root');
-// const theme = createTheme({
-//   palette: {
-//     primary: {
-//       main: '#1b1b1b',
-//     },
-//     secondary: {
-//       main: '#ffffff',
-//     },
-//     grey: {
-//       50: '#fafafa',
-//       100: '#f5f5f5',
-//     },
-//   },
-// });
+const theme = createTheme({
+  palette: {
+    primary: {
+      main: '#1b1b1b',
+    },
+    secondary: {
+      main: '#ffffff',
+    },
+    grey: {
+      50: '#fafafa',
+      100: '#f5f5f5',
+    },
+  },
+});
 
 if (container) {
   const root = createRoot(container);
   root.render(
-    // <ThemeProvider theme={theme}>
-    <Provider store={store}>
-      <ThemeProvider>
+    <ThemeProvider theme={theme}>
+      <Provider store={store}>
         <App />
-      </ThemeProvider>
-    </Provider>,
-    // </ThemeProvider>,
+      </Provider>
+    </ThemeProvider>,
   );
 }
